[🚀 feat] home and reload icons on navbar
