+++ components/icons/home.tsx
@@ -3,15 +3,15 @@ type Props = App.SVGProps;
 const Home: Nixix.FC<Props> = (props): someView => {
   return (
     <svg
-      width="258"
-      height="232"
+      width="190"
+      height="171"
       {...props}
-      viewBox="0 0 258 232"
+      viewBox="0 0 190 171"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        d="M0 114.934C0 120.218 4.18332 123.411 9.35774 123.411C12.5504 123.411 15.0822 121.87 17.2843 119.668L124.732 21.7978C125.943 20.5868 127.265 20.1462 128.696 20.1462C130.017 20.1462 131.228 20.5868 132.549 21.7978L239.888 119.668C242.199 121.87 244.731 123.411 247.813 123.411C252.987 123.411 257.172 120.218 257.172 114.934C257.172 111.632 255.961 109.54 253.869 107.669L215.777 72.9896V7.92651C215.777 3.0823 212.694 0 207.852 0H193.429C188.695 0 185.392 3.0823 185.392 7.92651V45.357L141.796 5.50464C137.943 1.87136 133.21 0.109915 128.586 0.109915C123.962 0.109915 119.338 1.87136 115.375 5.50464L3.30259 107.669C1.32132 109.54 0 111.632 0 114.934ZM31.4857 207.3C31.4857 222.603 40.7336 231.52 56.2564 231.52H100.843V153.246C100.843 148.181 104.256 144.879 109.32 144.879H148.182C153.246 144.879 156.549 148.181 156.549 153.246V231.52H201.024C216.548 231.52 225.686 222.603 225.686 207.3V126.714L132.219 42.4945C131.007 41.3935 129.687 40.8435 128.366 40.8435C127.154 40.8435 125.943 41.3935 124.623 42.6049L31.4857 127.264V207.3Z"
+        d="M41.6228 170.962H148.363C159.906 170.962 166.654 164.458 166.654 152.996V78.7742L177.14 88.3669C178.848 89.9929 180.718 91.131 182.994 91.131C186.815 91.131 189.903 88.7734 189.903 84.8712C189.903 82.4325 189.009 80.8879 187.466 79.5061L159.256 53.7353V5.8532C159.256 2.27607 156.98 0 153.403 0H142.753C139.257 0 136.819 2.27607 136.819 5.8532V33.3305L104.707 4.06481C101.862 1.46339 98.4475 0 94.9519 0C91.4563 0 88.0418 1.46339 85.1966 4.06481L2.43875 79.5061C0.975707 80.8879 0 82.4325 0 84.8712C0 88.7734 3.0891 91.131 6.91007 91.131C9.26765 91.131 11.1372 89.9929 12.7633 88.3669L23.2501 78.8554V152.996C23.2501 164.458 29.9975 170.962 41.6228 170.962ZM118.365 103.894C118.365 100.155 115.926 97.7972 112.187 97.7972H77.7173C73.9779 97.7972 71.458 100.155 71.458 103.894V157.874H44.3869C39.184 157.874 36.3387 155.028 36.3387 149.744V66.9051L92.1067 16.0962C92.9197 15.2832 93.895 14.7955 94.9519 14.7955C96.0088 14.7955 96.9842 15.202 97.8787 16.0962L153.565 66.9051V149.744C153.565 155.028 150.719 157.874 145.598 157.874H118.365V103.894Z"
         fill="inherit"
       />
     </svg>
