+++ components/icons/reload.tsx
@@ -1,11 +1,23 @@
-
-
 type Props = App.SVGProps;
 
 const Reload: Nixix.FC<Props> = (props): someView => {
   return (
-    <svg xmlns="http://www.w3.org/2000/svg" width={32} height={32} fill="none" {...props} viewBox="0 0 16 16"><path fill="inherit" fill-rule="evenodd" d="M8 1.5a6.5 6.5 0 1 0 6.445 7.348.75.75 0 1 0-1.487-.194A5.001 5.001 0 1 1 11.57 4.5h-1.32a.75.75 0 0 0 0 1.5h3a.75.75 0 0 0 .75-.75v-3a.75.75 0 0 0-1.5 0v1.06A6.48 6.48 0 0 0 8 1.5" clip-rule="evenodd" /></svg>
-  )
-}
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={32}
+      height={32}
+      fill="none"
+      {...props}
+      viewBox="0 0 16 16"
+    >
+      <path
+        fill="inherit"
+        fill-rule="evenodd"
+        d="M8 1.5a6.5 6.5 0 1 0 6.445 7.348.75.75 0 1 0-1.487-.194A5.001 5.001 0 1 1 11.57 4.5h-1.32a.75.75 0 0 0 0 1.5h3a.75.75 0 0 0 .75-.75v-3a.75.75 0 0 0-1.5 0v1.06A6.48 6.48 0 0 0 8 1.5"
+        clip-rule="evenodd"
+      />
+    </svg>
+  );
+};
 
 export default Reload;
