+++ src/frames/iphone-14/svg/device-frame.tsx
@@ -1,7 +1,6 @@
 import { concat } from "nixix/primitives";
 import { deviceFrameHeightClass } from "~/constants";
-import Iphone15ProImage from './device-image.png';
-
+import Iphone15Image from './device-image.png';
 interface Props extends App.SVGProps {
   height: number
 }
@@ -9,7 +8,7 @@ interface Props extends App.SVGProps {
 const DeviceFrame: Nixix.FC<Props> = ({ className, height, ...rest }): someView => {
 
   return (
-    <img src={Iphone15ProImage}
+    <img src={Iphone15Image}
       alt="Iphone Image"
       style={{
         ...rest.style,
