+++ src/frames/iphone-15/svg/device-frame.tsx
@@ -1,6 +1,7 @@
 import { concat } from "nixix/primitives";
 import { deviceFrameHeightClass } from "~/constants";
-import Iphone15Image from './device-image.png';
+import Iphone15ProImage from './device-image.png';
+
 interface Props extends App.SVGProps {
   height: number
 }
@@ -8,7 +9,7 @@ interface Props extends App.SVGProps {
 const DeviceFrame: Nixix.FC<Props> = ({ className, height, ...rest }): someView => {
 
   return (
-    <img src={Iphone15Image}
+    <img src={Iphone15ProImage}
       alt="Iphone Image"
       style={{
         ...rest.style,
