+++ src/frames/iphone-15/svg/status-bar.tsx
@@ -8,49 +8,48 @@ const StatusBar: Nixix.FC<StatusBarProps> = ({ className, foreground = '#030712'
   return (
     <svg
       {...rest} className={className + ' tws-h-fit'}
-      width="351" height="54" viewBox="0 0 351 54" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clip-path="url(#clip0_320_1899)">
-        <path d="M44.0166 25.7231C44.6585 25.7231 45.2673 25.8421 45.8428 26.0801C46.4183 26.3125 46.9274 26.6777 47.3701 27.1758C47.8184 27.6683 48.1698 28.3019 48.4243 29.0767C48.6789 29.8459 48.8062 30.7728 48.8062 31.8574V31.874C48.8062 33.2188 48.6125 34.3698 48.2251 35.3271C47.8377 36.2845 47.2816 37.0205 46.5566 37.5352C45.8372 38.0443 44.9767 38.2988 43.9751 38.2988C43.2391 38.2988 42.575 38.166 41.9829 37.9004C41.3963 37.6292 40.9093 37.2585 40.522 36.7881C40.1401 36.3122 39.8828 35.7699 39.75 35.1611L39.7334 35.0698H41.8335L41.8667 35.1528C41.9718 35.4295 42.1213 35.673 42.3149 35.8833C42.5142 36.0936 42.7521 36.2568 43.0288 36.373C43.311 36.4837 43.6265 36.5391 43.9751 36.5391C44.606 36.5391 45.1206 36.3537 45.519 35.9829C45.923 35.6066 46.2246 35.1058 46.4238 34.4805C46.6286 33.8496 46.7448 33.1523 46.7725 32.3887C46.778 32.3057 46.7808 32.2254 46.7808 32.1479C46.7808 32.0649 46.7808 31.9847 46.7808 31.9072L46.4072 29.9067C46.4072 29.453 46.3021 29.0407 46.0918 28.6699C45.8815 28.2992 45.5965 28.0059 45.2368 27.79C44.8826 27.5742 44.4814 27.4663 44.0332 27.4663C43.6016 27.4663 43.2059 27.5715 42.8462 27.7817C42.4865 27.992 42.1987 28.2798 41.9829 28.645C41.7671 29.0047 41.6592 29.4087 41.6592 29.8569V29.8735C41.6592 30.3384 41.7616 30.7507 41.9663 31.1104C42.1711 31.4645 42.4505 31.744 42.8047 31.9487C43.1589 32.1535 43.5628 32.2559 44.0166 32.2559C44.4704 32.2559 44.8771 32.1562 45.2368 31.957C45.5965 31.7578 45.8815 31.4811 46.0918 31.127C46.3021 30.7728 46.4072 30.3716 46.4072 29.9233V29.9067H46.9053V32.1147H46.6147C46.4709 32.4302 46.2578 32.7235 45.9756 32.9946C45.6934 33.2658 45.3447 33.4844 44.9297 33.6504C44.5202 33.8164 44.0387 33.8994 43.4854 33.8994C42.7217 33.8994 42.0465 33.7279 41.46 33.3848C40.8734 33.0417 40.4113 32.5713 40.0737 31.9736C39.7417 31.376 39.5757 30.7008 39.5757 29.9482V29.9316C39.5757 29.1182 39.7638 28.396 40.1401 27.7651C40.522 27.1287 41.0477 26.6307 41.7173 26.271C42.3924 25.9058 43.1589 25.7231 44.0166 25.7231ZM52.5747 36.3896C52.1984 36.3896 51.883 36.2651 51.6284 36.0161C51.3794 35.7616 51.2549 35.4517 51.2549 35.0864C51.2549 34.7157 51.3794 34.4058 51.6284 34.1567C51.883 33.9022 52.1984 33.7749 52.5747 33.7749C52.9565 33.7749 53.272 33.9022 53.521 34.1567C53.77 34.4058 53.8945 34.7157 53.8945 35.0864C53.8945 35.4517 53.77 35.7616 53.521 36.0161C53.272 36.2651 52.9565 36.3896 52.5747 36.3896ZM52.5747 30.2305C52.1984 30.2305 51.883 30.106 51.6284 29.8569C51.3794 29.6024 51.2549 29.2925 51.2549 28.9272C51.2549 28.5565 51.3794 28.2466 51.6284 27.9976C51.883 27.743 52.1984 27.6157 52.5747 27.6157C52.9565 27.6157 53.272 27.743 53.521 27.9976C53.77 28.2466 53.8945 28.5565 53.8945 28.9272C53.8945 29.2925 53.77 29.6024 53.521 29.8569C53.272 30.106 52.9565 30.2305 52.5747 30.2305ZM62.1538 38V35.7007H56.2935V33.9492C56.6089 33.4124 56.9243 32.8729 57.2397 32.3306C57.5607 31.7882 57.8844 31.2487 58.2109 30.7119C58.543 30.1696 58.8722 29.6356 59.1987 29.1099C59.5308 28.5786 59.86 28.0557 60.1865 27.541C60.5186 27.0208 60.8506 26.5145 61.1826 26.022H64.1958V33.9326H65.8145V35.7007H64.1958V38H62.1538ZM58.2856 33.9824H62.187V27.7319H62.0625C61.819 28.1082 61.5672 28.4956 61.3071 28.894C61.0526 29.2925 60.7952 29.6992 60.5352 30.1143C60.2751 30.5293 60.0177 30.9471 59.7632 31.3677C59.5086 31.7827 59.2568 32.2005 59.0078 32.6211C58.7588 33.0361 58.5181 33.4484 58.2856 33.8579V33.9824ZM70.77 38V28.1304H70.6206L67.6406 30.2305V28.2217L70.77 26.022H72.9033V38H70.77Z" fill="black" />
-        <path fill-rule="evenodd" clip-rule="evenodd" d="M269.615 27.033C269.615 26.4 269.137 25.8868 268.548 25.8868H267.482C266.893 25.8868 266.415 26.4 266.415 27.033V36.967C266.415 37.6 266.893 38.1132 267.482 38.1132H268.548C269.137 38.1132 269.615 37.6 269.615 36.967V27.033ZM262.181 28.3321H263.248C263.837 28.3321 264.314 28.8576 264.314 29.5058V36.9395C264.314 37.5877 263.837 38.1132 263.248 38.1132H262.181C261.592 38.1132 261.114 37.5877 261.114 36.9395V29.5058C261.114 28.8576 261.592 28.3321 262.181 28.3321ZM257.849 30.9811H256.782C256.193 30.9811 255.716 31.5133 255.716 32.1698V36.9245C255.716 37.581 256.193 38.1132 256.782 38.1132H257.849C258.438 38.1132 258.916 37.581 258.916 36.9245V32.1698C258.916 31.5133 258.438 30.9811 257.849 30.9811ZM252.548 33.4264H251.482C250.893 33.4264 250.415 33.951 250.415 34.5981V36.9415C250.415 37.5886 250.893 38.1132 251.482 38.1132H252.548C253.137 38.1132 253.615 37.5886 253.615 36.9415V34.5981C253.615 33.951 253.137 33.4264 252.548 33.4264Z" fill="black" />
-        <path fill-rule="evenodd" clip-rule="evenodd" d="M285.187 28.3022C287.674 28.3023 290.066 29.2243 291.868 30.8778C292.004 31.0055 292.221 31.0039 292.355 30.8742L293.652 29.6107C293.72 29.545 293.757 29.4559 293.757 29.3632C293.756 29.2705 293.718 29.1819 293.649 29.1169C288.918 24.7422 281.454 24.7422 276.723 29.1169C276.655 29.1819 276.616 29.2705 276.615 29.3632C276.615 29.4558 276.652 29.5449 276.72 29.6107L278.018 30.8742C278.151 31.0041 278.369 31.0057 278.504 30.8778C280.307 29.2242 282.699 28.3022 285.187 28.3022ZM285.183 32.5224C286.541 32.5223 287.849 33.0341 288.856 33.9582C288.992 34.0894 289.206 34.0865 289.339 33.9518L290.626 32.6325C290.694 32.5633 290.731 32.4694 290.73 32.3719C290.729 32.2743 290.69 32.1812 290.621 32.1134C287.557 29.2226 282.812 29.2226 279.748 32.1134C279.679 32.1812 279.64 32.2744 279.639 32.3719C279.638 32.4695 279.676 32.5634 279.743 32.6325L281.03 33.9518C281.163 34.0865 281.377 34.0894 281.513 33.9582C282.519 33.0347 283.827 32.523 285.183 32.5224ZM287.708 35.316C287.71 35.4213 287.672 35.5229 287.605 35.5967L285.429 38.0514C285.365 38.1236 285.278 38.1642 285.187 38.1642C285.096 38.1642 285.009 38.1236 284.945 38.0514L282.768 35.5967C282.701 35.5229 282.664 35.4212 282.666 35.3159C282.668 35.2105 282.709 35.1108 282.779 35.0402C284.169 33.7263 286.205 33.7263 287.595 35.0402C287.665 35.1108 287.706 35.2106 287.708 35.316Z" fill="black" />
-        <rect opacity="0.35" x="301.257" y="26" width="24" height="12" rx="3.8" stroke="black" />
-        <path opacity="0.4" d="M326.757 30.2811V34.3566C327.562 34.0114 328.085 33.2085 328.085 32.3189C328.085 31.4293 327.562 30.6263 326.757 30.2811" fill="black" />
-        <rect x="302.757" y="27.5" width="21" height="9" rx="2.5" fill="black" />
-        <g clip-path="url(#clip1_320_1899)">
-          <rect x="120" y="17" width="113" height="30.8626" rx="15.4313" fill="black" />
-          <path d="M216.697 38.1433C220.204 38.1433 223.047 35.4387 223.047 32.1024C223.047 28.7661 220.204 26.0615 216.697 26.0615C213.19 26.0615 210.347 28.7661 210.347 32.1024C210.347 35.4387 213.19 38.1433 216.697 38.1433Z" fill="#161616" />
-          <path d="M216.712 35.8279C218.875 35.8279 220.629 34.1596 220.629 32.1017C220.629 30.0439 218.875 28.3756 216.712 28.3756C214.549 28.3756 212.795 30.0439 212.795 32.1017C212.795 34.1596 214.549 35.8279 216.712 35.8279Z" fill="#0A0D13" />
-          <path d="M216.701 34.9409C218.35 34.9409 219.686 33.6698 219.686 32.1018C219.686 30.5339 218.35 29.2628 216.701 29.2628C215.053 29.2628 213.717 30.5339 213.717 32.1018C213.717 33.6698 215.053 34.9409 216.701 34.9409Z" fill="#091427" />
-          <g filter="url(#filter0_f_320_1899)">
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M214.657 30.2026C214.303 30.238 213.8 30.7519 213.763 31.7618C213.726 32.7718 214.191 33.3211 214.452 33.3211C214.713 33.3211 215.793 31.8504 214.657 30.2026Z" fill="#235A91" fill-opacity="0.556075" />
+      width="393" height="54" viewBox="0 0 393 54" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clip-path="url(#clip0_317_1609)">
+        <path d="M61.7666 25.7231C62.4085 25.7231 63.0173 25.8421 63.5928 26.0801C64.1683 26.3125 64.6774 26.6777 65.1201 27.1758C65.5684 27.6683 65.9198 28.3019 66.1743 29.0767C66.4289 29.8459 66.5562 30.7728 66.5562 31.8574V31.874C66.5562 33.2188 66.3625 34.3698 65.9751 35.3271C65.5877 36.2845 65.0316 37.0205 64.3066 37.5352C63.5872 38.0443 62.7267 38.2988 61.7251 38.2988C60.9891 38.2988 60.325 38.166 59.7329 37.9004C59.1463 37.6292 58.6593 37.2585 58.272 36.7881C57.8901 36.3122 57.6328 35.7699 57.5 35.1611L57.4834 35.0698H59.5835L59.6167 35.1528C59.7218 35.4295 59.8713 35.673 60.0649 35.8833C60.2642 36.0936 60.5021 36.2568 60.7788 36.373C61.061 36.4837 61.3765 36.5391 61.7251 36.5391C62.356 36.5391 62.8706 36.3537 63.269 35.9829C63.673 35.6066 63.9746 35.1058 64.1738 34.4805C64.3786 33.8496 64.4948 33.1523 64.5225 32.3887C64.528 32.3057 64.5308 32.2254 64.5308 32.1479C64.5308 32.0649 64.5308 31.9847 64.5308 31.9072L64.1572 29.9067C64.1572 29.453 64.0521 29.0407 63.8418 28.6699C63.6315 28.2992 63.3465 28.0059 62.9868 27.79C62.6326 27.5742 62.2314 27.4663 61.7832 27.4663C61.3516 27.4663 60.9559 27.5715 60.5962 27.7817C60.2365 27.992 59.9487 28.2798 59.7329 28.645C59.5171 29.0047 59.4092 29.4087 59.4092 29.8569V29.8735C59.4092 30.3384 59.5116 30.7507 59.7163 31.1104C59.9211 31.4645 60.2005 31.744 60.5547 31.9487C60.9089 32.1535 61.3128 32.2559 61.7666 32.2559C62.2204 32.2559 62.6271 32.1562 62.9868 31.957C63.3465 31.7578 63.6315 31.4811 63.8418 31.127C64.0521 30.7728 64.1572 30.3716 64.1572 29.9233V29.9067H64.6553V32.1147H64.3647C64.2209 32.4302 64.0078 32.7235 63.7256 32.9946C63.4434 33.2658 63.0947 33.4844 62.6797 33.6504C62.2702 33.8164 61.7887 33.8994 61.2354 33.8994C60.4717 33.8994 59.7965 33.7279 59.21 33.3848C58.6234 33.0417 58.1613 32.5713 57.8237 31.9736C57.4917 31.376 57.3257 30.7008 57.3257 29.9482V29.9316C57.3257 29.1182 57.5138 28.396 57.8901 27.7651C58.272 27.1287 58.7977 26.6307 59.4673 26.271C60.1424 25.9058 60.9089 25.7231 61.7666 25.7231ZM70.3247 36.3896C69.9484 36.3896 69.633 36.2651 69.3784 36.0161C69.1294 35.7616 69.0049 35.4517 69.0049 35.0864C69.0049 34.7157 69.1294 34.4058 69.3784 34.1567C69.633 33.9022 69.9484 33.7749 70.3247 33.7749C70.7065 33.7749 71.022 33.9022 71.271 34.1567C71.52 34.4058 71.6445 34.7157 71.6445 35.0864C71.6445 35.4517 71.52 35.7616 71.271 36.0161C71.022 36.2651 70.7065 36.3896 70.3247 36.3896ZM70.3247 30.2305C69.9484 30.2305 69.633 30.106 69.3784 29.8569C69.1294 29.6024 69.0049 29.2925 69.0049 28.9272C69.0049 28.5565 69.1294 28.2466 69.3784 27.9976C69.633 27.743 69.9484 27.6157 70.3247 27.6157C70.7065 27.6157 71.022 27.743 71.271 27.9976C71.52 28.2466 71.6445 28.5565 71.6445 28.9272C71.6445 29.2925 71.52 29.6024 71.271 29.8569C71.022 30.106 70.7065 30.2305 70.3247 30.2305ZM79.9038 38V35.7007H74.0435V33.9492C74.3589 33.4124 74.6743 32.8729 74.9897 32.3306C75.3107 31.7882 75.6344 31.2487 75.9609 30.7119C76.293 30.1696 76.6222 29.6356 76.9487 29.1099C77.2808 28.5786 77.61 28.0557 77.9365 27.541C78.2686 27.0208 78.6006 26.5145 78.9326 26.022H81.9458V33.9326H83.5645V35.7007H81.9458V38H79.9038ZM76.0356 33.9824H79.937V27.7319H79.8125C79.569 28.1082 79.3172 28.4956 79.0571 28.894C78.8026 29.2925 78.5452 29.6992 78.2852 30.1143C78.0251 30.5293 77.7677 30.9471 77.5132 31.3677C77.2586 31.7827 77.0068 32.2005 76.7578 32.6211C76.5088 33.0361 76.2681 33.4484 76.0356 33.8579V33.9824ZM88.52 38V28.1304H88.3706L85.3906 30.2305V28.2217L88.52 26.022H90.6533V38H88.52Z" fill="black" />
+        <path fill-rule="evenodd" clip-rule="evenodd" d="M307.865 27.033C307.865 26.4 307.387 25.8868 306.798 25.8868H305.732C305.143 25.8868 304.665 26.4 304.665 27.033V36.967C304.665 37.6 305.143 38.1132 305.732 38.1132H306.798C307.387 38.1132 307.865 37.6 307.865 36.967V27.033ZM300.431 28.3321H301.498C302.087 28.3321 302.564 28.8576 302.564 29.5058V36.9395C302.564 37.5877 302.087 38.1132 301.498 38.1132H300.431C299.842 38.1132 299.364 37.5877 299.364 36.9395V29.5058C299.364 28.8576 299.842 28.3321 300.431 28.3321ZM296.099 30.9811H295.032C294.443 30.9811 293.966 31.5133 293.966 32.1698V36.9245C293.966 37.581 294.443 38.1132 295.032 38.1132H296.099C296.688 38.1132 297.166 37.581 297.166 36.9245V32.1698C297.166 31.5133 296.688 30.9811 296.099 30.9811ZM290.798 33.4264H289.732C289.143 33.4264 288.665 33.951 288.665 34.5981V36.9415C288.665 37.5886 289.143 38.1132 289.732 38.1132H290.798C291.387 38.1132 291.865 37.5886 291.865 36.9415V34.5981C291.865 33.951 291.387 33.4264 290.798 33.4264Z" fill="black" />
+        <path fill-rule="evenodd" clip-rule="evenodd" d="M323.437 28.3021C325.924 28.3022 328.316 29.2243 330.118 30.8778C330.254 31.0054 330.471 31.0038 330.605 30.8742L331.902 29.6107C331.97 29.545 332.007 29.4559 332.007 29.3632C332.006 29.2705 331.968 29.1819 331.899 29.1169C327.168 24.7422 319.704 24.7422 314.973 29.1169C314.905 29.1818 314.866 29.2704 314.865 29.3631C314.865 29.4558 314.902 29.5449 314.97 29.6107L316.268 30.8742C316.401 31.004 316.619 31.0056 316.754 30.8778C318.557 29.2242 320.949 28.3021 323.437 28.3021ZM323.433 32.5224C324.791 32.5223 326.099 33.0341 327.106 33.9582C327.242 34.0893 327.456 34.0865 327.589 33.9518L328.876 32.6325C328.944 32.5633 328.981 32.4694 328.98 32.3718C328.979 32.2743 328.94 32.1812 328.871 32.1134C325.807 29.2226 321.062 29.2226 317.998 32.1134C317.929 32.1812 317.89 32.2743 317.889 32.3719C317.888 32.4695 317.926 32.5634 317.993 32.6325L319.28 33.9518C319.413 34.0865 319.627 34.0893 319.763 33.9582C320.769 33.0347 322.077 32.523 323.433 32.5224ZM325.958 35.316C325.96 35.4213 325.922 35.5229 325.855 35.5967L323.679 38.0514C323.615 38.1236 323.528 38.1642 323.437 38.1642C323.346 38.1642 323.259 38.1236 323.195 38.0514L321.018 35.5967C320.951 35.5228 320.914 35.4212 320.916 35.3159C320.918 35.2105 320.959 35.1108 321.029 35.0401C322.419 33.7262 324.455 33.7262 325.845 35.0401C325.915 35.1108 325.956 35.2106 325.958 35.316Z" fill="black" />
+        <rect opacity="0.35" x="339.507" y="26" width="24" height="12" rx="3.8" stroke="black" />
+        <path opacity="0.4" d="M365.007 30.2811V34.3566C365.812 34.0114 366.335 33.2085 366.335 32.3189C366.335 31.4293 365.812 30.6263 365.007 30.2811Z" fill="black" />
+        <rect x="341.007" y="27.5" width="21" height="9" rx="2.5" fill="black" />
+        <g clip-path="url(#clip1_317_1609)">
+          <rect x="139" y="15" width="127" height="35.3113" rx="17.6556" fill="black" />
+          <path d="M247.347 39.1909C251.36 39.1909 254.613 36.0964 254.613 32.2792C254.613 28.4619 251.36 25.3675 247.347 25.3675C243.334 25.3675 240.081 28.4619 240.081 32.2792C240.081 36.0964 243.334 39.1909 247.347 39.1909Z" fill="#161616" />
+          <path d="M247.36 36.5417C249.835 36.5417 251.841 34.633 251.841 32.2784C251.841 29.9239 249.835 28.0152 247.36 28.0152C244.884 28.0152 242.878 29.9239 242.878 32.2784C242.878 34.633 244.884 36.5417 247.36 36.5417Z" fill="#0A0D13" />
+          <path d="M247.35 35.5268C249.236 35.5268 250.765 34.0725 250.765 32.2786C250.765 30.4846 249.236 29.0303 247.35 29.0303C245.464 29.0303 243.936 30.4846 243.936 32.2786C243.936 34.0725 245.464 35.5268 247.35 35.5268Z" fill="#091427" />
+          <g filter="url(#filter0_f_317_1609)">
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M245.006 30.1055C244.601 30.1461 244.026 30.734 243.983 31.8895C243.94 33.0451 244.473 33.6736 244.771 33.6736C245.07 33.6736 246.306 31.9909 245.006 30.1055Z" fill="#235A91" fill-opacity="0.556075" />
           </g>
-          <g filter="url(#filter1_f_320_1899)">
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M217.189 29.8491C216.63 30.0125 216.523 30.8703 216.545 31.3605C216.566 31.8507 217.296 33.1375 218.176 32.8515C219.057 32.5656 219.207 31.5035 218.842 30.7887C218.477 30.0738 217.811 29.5632 217.189 29.8491Z" fill="#235A91" fill-opacity="0.556075" />
+          <g filter="url(#filter1_f_317_1609)">
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M247.907 29.7011C247.268 29.8881 247.145 30.8696 247.17 31.4304C247.194 31.9913 248.03 33.4635 249.037 33.1363C250.044 32.8092 250.216 31.594 249.798 30.7761C249.381 29.9582 248.619 29.374 247.907 29.7011Z" fill="#235A91" fill-opacity="0.556075" />
           </g>
         </g>
       </g>
       <defs>
-        <filter id="filter0_f_320_1899" x="213.608" y="30.0494" width="1.7126" height="3.42488" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
+        <filter id="filter0_f_317_1609" x="243.805" y="29.9303" width="1.9589" height="3.91853" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
           <feFlood flood-opacity="0" result="BackgroundImageFix" />
           <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.0765867" result="effect1_foregroundBlur_320_1899" />
+          <feGaussianBlur stdDeviation="0.0876262" result="effect1_foregroundBlur_317_1609" />
         </filter>
-        <filter id="filter1_f_320_1899" x="216.389" y="29.6151" width="2.7917" height="3.43055" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
+        <filter id="filter1_f_317_1609" x="246.992" y="29.4334" width="3.19425" height="3.92506" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
           <feFlood flood-opacity="0" result="BackgroundImageFix" />
           <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.0765867" result="effect1_foregroundBlur_320_1899" />
+          <feGaussianBlur stdDeviation="0.0876262" result="effect1_foregroundBlur_317_1609" />
         </filter>
-        <clipPath id="clip0_320_1899">
-          <rect width="351" height="54" fill="white" />
+        <clipPath id="clip0_317_1609">
+          <rect width="393" height="54" fill="white" />
         </clipPath>
-        <clipPath id="clip1_320_1899">
-          <rect width="112.93" height="30.861" fill="white" transform="translate(120 17)" />
+        <clipPath id="clip1_317_1609">
+          <rect width="127" height="35.3095" fill="white" transform="translate(139 15)" />
         </clipPath>
       </defs>
     </svg>
-
   )
-}
+};
 
 
 export default StatusBar;
